requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.3.2
aiohttp==3.10.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==4.3.0
async-cache==1.1.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
black==24.4.2
bs4==0.0.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dataclass-factory==2.16
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fake-headers==1.0.2
fastapi==0.111.1
fastapi-cli==0.0.4
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
motor==3.5.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.6
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pyasn1==0.6.0
pycparser==2.22
pydantic==1.10.17
pydub==0.25.1
Pygments==2.18.0
pymongo==4.8.0
pytest==8.3.2
pytest-asyncio==0.23.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
rsa==4.9
shazamio==0.6.0
shazamio_core==1.0.7
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
webencodings==0.5.1
websockets==12.0
yarl==1.9.4
